+++ src/pages/user/MyBooksPage/ToRead.jsx
@@ -10,7 +10,7 @@ const ToRead = () => {
 
     return (
         <>
-            {booksDisplayed(books)}
+            {books && booksDisplayed(books)}
         </>
     )
 }
